Change structure and add .env option
