+++ .gitignore
@@ -1,2 +1,4 @@
 tmp
 go.sum
+.env
+
