+++ config.go
@@ -4,26 +4,9 @@ import (
 	"fmt"
 	"log"
 	"os"
+
+	"github.com/joho/godotenv"
 )
-
-func readEnvWithDefault(key string, defaultValue string) string {
-	value := os.Getenv(key)
-	if value == "" {
-		value = defaultValue
-		log.Printf("%s environment variable not set, using default: %s", key, defaultValue)
-	}
-
-	return value
-}
-
-func readEnv(key string) (string, error) {
-	value := os.Getenv(key)
-	if value == "" {
-		return "", fmt.Errorf("%s environment variable is not defined", key)
-	}
-
-	return value, nil
-}
 
 type ServerConfig struct {
 	FtpServerURL  string
@@ -34,18 +17,16 @@ type ServerConfig struct {
 
 var config *ServerConfig
 
-func init() {
-	loadConfig()
-}
+func LoadConfig() {
+	godotenv.Load()
 
-func loadConfig() {
-	FtpServerURL, err := readEnv("FTP_SERVER_URL")
+	FtpServerURL, err := ReadEnv("FTP_SERVER_URL")
 	if err != nil {
 		panic(err)
 	}
-	FtpServerPath := readEnvWithDefault("FTP_SERVER_PATH", "/")
-	FtpUsername := readEnvWithDefault("FTP_USERNAME", "ftpuser")
-	FtpPassword := readEnvWithDefault("FTP_PASSWORD", "ftpuser")
+	FtpServerPath := ReadEnvWithDefault("FTP_SERVER_PATH", "/")
+	FtpUsername := ReadEnvWithDefault("FTP_USERNAME", "ftpuser")
+	FtpPassword := ReadEnvWithDefault("FTP_PASSWORD", "ftpuser")
 
 	config = &ServerConfig{
 		FtpServerURL:  FtpServerURL,
@@ -53,4 +34,25 @@ func loadConfig() {
 		FtpUsername:   FtpUsername,
 		FtpPassword:   FtpPassword,
 	}
+
+	fmt.Printf("** Server is loaded with the following: %s ** \n", *config)
+}
+
+func ReadEnvWithDefault(key string, defaultValue string) string {
+	value := os.Getenv(key)
+	if value == "" {
+		value = defaultValue
+		log.Printf("%s environment variable not set, using default: %s", key, defaultValue)
+	}
+
+	return value
+}
+
+func ReadEnv(key string) (string, error) {
+	value := os.Getenv(key)
+	if value == "" {
+		return "", fmt.Errorf("%s environment variable is not defined", key)
+	}
+
+	return value, nil
 }
