+++ main.go
@@ -4,11 +4,10 @@ import (
 	"github.com/gin-gonic/gin"
 )
 
-func healthEndpoint(c *gin.Context) {
-	c.JSON(200, gin.H{
-		"message": "pong",
-	})
+func init() {
+	LoadConfig()
 }
+
 func main() {
 	server := gin.Default()
 	v1 := server.Group("/v1")
@@ -16,5 +15,6 @@ func main() {
 		v1.GET("/ping", healthEndpoint)
 		v1.GET("/list", ftpListEndpoint)
 	}
+
 	server.Run()
 }
